refactor: Migrate rdbms properties to unified configuration
